requirements.txt
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.1
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.2.0
future==0.18.3
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pydantic==2.6.0
pydantic_core==2.16.1
pydub==0.25.1
Pygments==2.17.2
pynput==1.7.6
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rsa==4.9
safetensors==0.4.2
six==1.16.0
SpeechRecognition==3.10.1
sympy==1.12
tdqm==0.0.1
tiktoken==0.5.2
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
urllib3==2.2.0
whisper_mic==1.4.0
zipp==3.17.0
